switch to connect modal in useConnect hook; wrap storybook in context providers
