--- packages/react/.storybook/preview.ts
@@ -1,14 +1,0 @@
-import "../src/globals.css";
-
-const preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
