+++ packages/react/src/i18n.ts
@@ -1,0 +1,15 @@
+import i18n from "i18next";
+import { initReactI18next } from "react-i18next";
+
+import { en } from "./locales/en";
+
+i18n.use(initReactI18next).init({
+  resources: { en },
+  lng: "en",
+  fallbackLng: "en",
+  interpolation: {
+    escapeValue: false,
+  },
+});
+
+export { i18n };
