+++ packages/react/src/index.ts
@@ -1,8 +1,4 @@
-import i18n from "i18next";
-import { initReactI18next } from "react-i18next";
-
 import "./globals.css";
-import { en } from "./locales/en";
 
 export * from "@treasure-dev/tdk-core";
 export { ConnectButton } from "./components/connect/ConnectButton";
@@ -17,12 +13,3 @@ export {
   useContractAddress,
   useContractAddresses,
 } from "./hooks/useContractAddress";
-
-i18n.use(initReactI18next).init({
-  resources: { en },
-  lng: "en",
-  fallbackLng: "en",
-  interpolation: {
-    escapeValue: false,
-  },
-});
